+++ lib/server/ProtocolUncertainStream.h
@@ -31,15 +31,15 @@ private:
 	ProtocolUncertainStream(const ProtocolUncertainStream &);
 	
 public:
-	virtual int Read(void *pBuffer, int NBytes, int Timeout = IOStream::TimeOutInfinite);
+	virtual size_t Read(void *pBuffer, size_t NBytes, int Timeout = IOStream::TimeOutInfinite);
 	virtual pos_type BytesLeftToRead();
-	virtual void Write(const void *pBuffer, int NBytes);
+	virtual void Write(const void *pBuffer, size_t NBytes);
 	virtual bool StreamDataLeft();
 	virtual bool StreamClosed();
 
 private:
 	IOStream &mrSource;
-	int mBytesLeftInCurrentBlock;
+	size_t mBytesLeftInCurrentBlock;
 	bool mFinished;
 };
